# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Doc IT API
  description: Backend API for Doc IT application
  version: 1.0.0
  contact:
    email: support@docit.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.docit.com/api/v1
    description: Production server
...
paths:
...
  /documents/{id}:
    get:
      tags: [Documents]
      summary: Get a specific document by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Document ID
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

    put:
      tags: [Documents]
      summary: Update a document by ID
      description: Updates an existing document's name and/or files. All fields except ID are optional.
      operationId: updateDocumentById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the document to update (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name1:
                  type: string
                  description: Name of the first document (required if provided, must be non-empty)
                  example: UpdatedFirstDocument
                name2:
                  type: string
                  description: Name of the second document (optional, can be empty or omitted)
                  example: UpdatedSecondDocument
                document:
                  type: string
                  format: binary
                  description: 'First document file (optional, supported types: pdf, doc, docx, txt)'
                document2:
                  type: string
                  format: binary
                  description: 'Second document file (optional, supported types: pdf, doc, docx, txt)'
              required: []
      responses:
        '200':
          description: Document updated successfully

    delete:
      tags: [Documents]
      summary: Delete a document
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Document ID
      responses:
        '200':
          description: Document deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
          additionalProperties: true
          example:
            _id: 507f1f77bcf86cd799439011
            name: document.pdf
            filePath: /Uploads/document.pdf
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT